fix: 🐛 fix layout demo error in codesanbox
