+++ sites/x6-sites-demos/packages/tutorial/advanced/layout/dagre/src/app.css
@@ -7,7 +7,7 @@
 
 .app-content {
   flex: 1;
-  height: 300px;
+  height: 400px;
   margin-left: 8px;
   margin-right: 8px;
   box-shadow: 0 0 10px 1px #e9e9e9;
